+++ dataforge-data/src/commonMain/kotlin/space/kscience/dataforge/actions/SplitAction.kt
@@ -87,7 +87,6 @@ internal class SplitAction<T : Any, R : Any>(
  * Action that splits each incoming element into a number of fragments defined in builder
  */
 @DFExperimental
-@Suppress("FunctionName")
 public inline fun <T : Any, reified R : Any> Action.Companion.split(
     noinline builder: SplitBuilder<T, R>.() -> Unit,
 ): Action<T, R> = SplitAction(typeOf<R>(), builder)