+++ CHANGELOG.md
@@ -14,6 +14,12 @@
 
 ### Security
 
+## 0.6.2 - 2023-07-29
+
+### Changed
+- Meta to Json serializer now serializes a single item with index as an array. It is important for plotly integration.
+- Meta to Json serializes Meta without children a value as literal or array instead of an object with `@value` field.
+
 ## 0.6.1 - 2023-03-31
 
 ### Added